% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-element.R
\name{add_abstract}
\alias{add_abstract}
\title{Add Abstract}
\usage{
add_abstract(parent_element, abstract_file)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{abstract_file}{A file containing abstract information (word or markdown): see \code{\link{create_abstract}}}
}
\description{
Add Abstract
}
