% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{add_attribute}
\alias{add_attribute}
\title{Add Atribute}
\usage{
add_attribute(
  attribute_name,
  attribute_label = NULL,
  attribute_definition,
  storage_type,
  measurement_scale,
  text_definition = NULL,
  text_pattern = NULL,
  domain = NULL,
  units = NULL,
  number_type = NULL,
  code_number = NULL,
  code_number_definition = NULL,
  unit_precision = NULL,
  date_time_format = NULL,
  date_time_precision = NULL,
  minimum = NULL,
  maximum = NULL
)
}
\arguments{
\item{attribute_name}{The name    of a field in    a    data table. This is    often    a short    and/or cryptic name. It is recommended that the attribute names    be suitable    for    use    as a    variable,    e.g.,    composed of ASCII    characters,    and    that the    attribute names    match    the    column headers of    a    CSV    or other text table.}

\item{attribute_label}{Optional. Used to provide    a    less ambiguous or    less cryptic alternative identification    than what    is provided    in attribute_name}

\item{storage_type}{An indicator    to processing    systems    as to    how    the    attribute    might    be represented in    a    system or    language,    but    is distinct    from the actual    expression of    the    domain of    the    attribute. Non system-specific examples    include: float,    integer    and    string.}

\item{measurement_scale}{The    type of    scale    from which values are    drawn    for    the    attribute. Must use one of the following types: nominal, ordinal, interval, ratio, or dateTime.}

\item{text_definition}{Description for nominal measurement scale values.}

\item{text_pattern}{TODO}

\item{domain}{TODO}

\item{units}{The units assigned to this particular attributes values.}

\item{number_type}{Can be defined as: real, natural, whole, or integer}

\item{code_number}{The reference number to the code definition.}

\item{code_number_definition}{The explaination of a listed code number}

\item{unit_precision}{How precise units are measured}

\item{date_time_format}{YYYY-MM-DDThh:mm:ssTZD format}

\item{date_time_precision}{What level of time is being measured.}

\item{minimum}{Theoretical or allowable minimum value.}

\item{maximum}{Theoretical or allowable maximum value.}

\item{attribute_defenition}{A precise and complete    definition of the attribute    being    documented.}
}
\value{
The project or dataset list with an attribute list appended
}
\description{
adds an attribute element according to EML standards. Multiple attributes can be created. Will be appended to the attribute list.
}
\section{Measurement Scales}{

Different measurement scale values will indicate different inputs:

\subsection{Non-numeric}
Please provide a domain to indicate if your attribute is text or enumerated.
\subsubsection{Nominal:}
Used to define nominal (categorical) scale attributes. If your attribute falls
under the domain of text, please provide the inputs of text_definition and text_pattern.
If your attribute falls under the domain of enumerated, please provide the inputs
of code number and code number definition.
\subsubsection{ordinal}
1.Nominal: Please provide measurement_scale_definition.
2.Ordinal: Please provide code_number and code_number_definition.
3.Interval: Please provide units, unit_precision, number_type, minimum, and maximum.
4.Ratio: Please provide units, unit_precision, and number_type.
5.dateTime: Please provide ISO_date_time, date_time_precision, minimum, and maximum
}

\examples{
Nominal:
add_attribute(attribute_name = "site_id", attribute_definition = "Site id as used in sites table",
                   storage_type = "typeSystem = 'http://www.w3.org/2001/XMLSchema-datatypes'>string<",
                   measurement_scale = "nominal", nominal_scale_definition = "Site id as used in sites table.")

Ordinal:
add_attribute(attribute_name = "q110", attribute_definition = "Q110 - preference for front yard",
              storage_type = "typeSystem = 'http://www.w3.org/2001/XMLSchema-datatypes'>float<",
              measurement_scale = "ordinal", code_number = c("1.00", "2.00", "3.00"),
              code_number_definition = c("A desert landscape.", "Mostly lawn.", "Some lawn.")
 
Interval:
add_attribute(attribute_name = "Count", attribute_definition = "Number of individuals observed",
                   measurement_scale = "interval",  storage_type = "integer", units = "number", unit_precision = "1",
                   number_type = "whole", minimum = "0")
                   
Ratio: 
add_attribute(attribute_name = "pH", attribute_definition = "pH of soil solution",
                   storage_type = "typeSystem = 'http://www.w3.org/2001/XMLSchema-datatypes'>float<",
                   measurement_scale = "ratio", units = "dimensionless", unit_precision = "0.01",
                   number_type = "real")
                   
dateTime:
add_attribute(attribute_name = "Yrs", attribute_label = "Years",
                   attribute_definition = "Calendar year of the observation from years 1990 - 2010.",
                   storage_type = "integer", measurement_scale = "dateTime", ISO_date_time = "YYYY",
                   date_time_precision = "1", minimum = "1993", maximum = "2003")
}
