% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage-element.R
\name{add_coverage}
\alias{add_coverage}
\title{Add Coverage Element}
\usage{
add_coverage(
  parent_element,
  geographic_description,
  west_bounding_coordinate,
  east_bounding_coordinate,
  north_bounding_coordinate,
  south_bounding_coordinate,
  begin_date,
  end_date
)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset}

\item{geographic_description}{#TODO}

\item{west_bounding_coordinate}{#TODO}

\item{east_bounding_coordinate}{#TODO}

\item{north_bounding_coordinate}{#TODO}

\item{south_bounding_coordinate}{#TODO}

\item{begin_date}{#TODO}

\item{end_date}{#TODO}
}
\value{
The dataset or project with coverage information appended
}
\description{
Adds coverage element #TODO
}
\examples{
#TODO
}
