% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-table-element .R
\name{add_data_table}
\alias{add_data_table}
\title{Add Data Table}
\usage{
add_data_table(
  parent_element,
  alternate_identifier = NULL,
  entity_name,
  entity_description,
  object_name,
  number_of_headers,
  record_delimiter,
  physcial_line_delimiter,
  attribute_orientation,
  field_delimiter,
  online_url,
  attribute_list
)
}
\arguments{
\item{entity_name}{TODO}

\item{entity_description}{TODO}

\item{object_name}{TODO}

\item{number_of_headers}{TODO}

\item{record_delimiter}{TODO}

\item{attribute_orientation}{TODO}

\item{field_delimiter}{TODO}

\item{online_url}{TODO}

\item{attribute_list}{TODO}

\item{parament_element}{a list representing the EML project or dataset}

\item{alternate_identifer}{TODO}

\item{physical_line_delimiter}{TODO}
}
\value{
the project or dataset list with a data table appended
}
\description{
adds the data table elements according to EML standards
}
\examples{
TODO
}
