% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-table-element.R
\name{add_datatable}
\alias{add_datatable}
\title{Add Data Table}
\usage{
add_datatable(parent_element, datatable_metadata)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{datatable_metadata}{A named list or dataframe containing datatable metadata elements
(filepath, attribute_info, datatable_description, datatable_url): see \code{\link{create_datatable}}}
}
\value{
The dataset list or project with datatable information appended.
}
\description{
Adds the data table elements to a dataset list according to EML standards.
}
\examples{
datatable_metadata <- dplyr::tibble(filepath = c(system.file("extdata", "Banet-Example", "data", "enclosure-study-growth-rate-data.csv", package = "EMLaide", mustWork = TRUE), 
                                                             system.file("extdata", "Banet-Example", "data", "enclosure-study-gut-contents-data.csv", package = "EMLaide", mustWork = TRUE)), 
                                     attribute_info = c(system.file("extdata", "Banet-Example", "metadata", "enclosure-study-growth-rates-metadata.xlsx", package = "EMLaide", mustWork = TRUE), 
                                     system.file("extdata", "Banet-Example", "metadata", "enclosure-study-gut-contents-metadata.xlsx", package = "EMLaide", mustWork = TRUE)),
                                     datatable_description = c("Growth Rates - Enclosure Study","Gut Contents - Enclosure Study"))
 
dataset <- list() \%>\%
    add_datatable(datatable_metadata)
dataset
}
