% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{add_datetime}
\alias{add_datetime}
\title{Add dateTime Measurement Scale}
\usage{
add_datetime(date_time_format, date_time_precision, minimum, maximum)
}
\arguments{
\item{date_time_format}{ISO 8601 format should be used.}

\item{date_time_precision}{To what level of time your attribute is recorded.}

\item{minimum}{The earliest dateTime recorded.}

\item{maximum}{The latest dateTime recorded.}
}
\description{
Add dateTime Measurement Scale
}
\keyword{internal}
