% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding-element.R
\name{add_funding}
\alias{add_funding}
\title{Add Funding Element}
\usage{
add_funding(
  parent_element,
  funder_name,
  funder_identifier,
  award_number,
  award_title,
  award_url,
  funding_description
)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset}

\item{funder_name}{Organization or individual providing the funding}

\item{funder_identifier}{This is where the funding organization is listed in the registry. The funder identifier must be registered. Follow the instructions at
\url{https://ror.org/curation/} to get registered.}

\item{award_number}{The identifier assigned by the funding agency to identify this funding award.}

\item{award_title}{Title of the dataset or project which recieved funding}

\item{award_url}{A link to information about the funding award on the funding organization's webpage.}

\item{funding_description}{A short description of the funding recieved.}
}
\value{
The dataset or project with funding information appended
}
\description{
Adds the funding information of a dataset based off of EML standards
}
\examples{
add_funding(parent_element = list(), funder_name = "National Science Foundation",
            funder_identifier = "http://dx.doi.org/10.13039/100000001",
            award_number = "1656026",
            award_title = "LTER: Beaufort Sea Lagoons: An Arctic Coastal Ecosystem in Transition",
            award_url = "https://www.nsf.gov/awardsearch/showAward?AWD_ID=1656026",
            funding_description = "BLE LTER is supported by the National Science Foundation under award #1656026 (2017-08-01 to 2022-07-31)." )
}
