% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{add_interval}
\alias{add_interval}
\title{Add Interval Measurement Scale}
\usage{
add_interval(
  units,
  unit_precision,
  number_type,
  minimum = NULL,
  maximum = NULL
)
}
\arguments{
\item{units}{The units assigned to this attribute.}

\item{unit_precision}{How precise this attirbutes' measurements are recorded.}

\item{number_type}{What type of number. Examples given in exported documentation.}

\item{minimum}{Optional. A theoreical minimum.}

\item{maximum}{Optional. A theoretical maximum.}
}
\description{
Add Interval Measurement Scale
}
\keyword{internal}
