% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{add_interval_ratio}
\alias{add_interval_ratio}
\title{Add Interval or Ratio Measurement Scales}
\usage{
add_interval_ratio(
  units,
  unit_precision,
  number_type,
  minimum = NULL,
  maximum = NULL
)
}
\arguments{
\item{units}{The units assigned to this attribute.}

\item{unit_precision}{How precise this attirbutes' measurements are recorded.}

\item{number_type}{What type of number. Examples given in exported documentation.}

\item{minimum}{Optional. A theoreical minimum.}

\item{maximum}{Optional. A theoretical maximum.}

\item{type}{Either "interval" or "ratio". Use "interval" to define
data which consist of equidistant points on a scale. Use "ratio" to define data
which consists not only of equidistant points but also has a meaningful zero
point, which allows the ratio to have meaning.}
}
\description{
Add Interval or Ratio Measurement Scales
}
\keyword{internal}
