% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword-set-element.R
\name{add_keyword_set}
\alias{add_keyword_set}
\title{Add Keywords}
\usage{
add_keyword_set(parent_element, keyword_metadata)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{keyword_metadata}{A named list or dataframe containing keyword elements: see \code{\link{create_keyword_set}}}
}
\value{
The dataset list or project with keyword information appended.
}
\description{
Adds the keyword metadata elements to a dataset list according to EML standards.
}
\examples{
keyword_metadata <- dplyr::tibble(keyword = c("Sacramento River", "Salmonid Habitat Restoration Projects", "Effectiveness Monitoring", "Pacific Salmon", "CVPIA"), 
                                  keywordThesaurus = c(NA, NA, NA, NA, NA))

dataset <- list() \%>\%
     add_keyword_set(keyword_metadata)
dataset
}
