% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intellectual-rights-element.R
\name{add_license}
\alias{add_license}
\title{Add License}
\usage{
add_license(parent_element, license_metadata)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{license_metadata}{A table or list containing license metadata: see \code{\link{create_license}}}
}
\value{
The dataset list or project with intellectual rights information appended.
}
\description{
Adds the license element to a dataset list according to EML standards.
}
\examples{
license_metadata = list(default_license = "CCBY")
dataset <- list() \%>\%
    add_license(license_metadata)
dataset
}
