% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{add_ordinal}
\alias{add_ordinal}
\title{Add Ordinal Measurment Scale}
\usage{
add_ordinal(domain = c("text", "enumerated"), definition, text_pattern = NULL)
}
\arguments{
\item{domain}{Either "text" or "enumerated".}

\item{definition}{If domain is "text", provide a word description of your attribute.
If the domain is "enumerated", provide a list of the attributes code definitions.
Examples are seen in exported documentation.}

\item{text_pattern}{Optional. Only needed if applicable and the domain given is "text".}
}
\description{
Add Ordinal Measurment Scale
}
\keyword{internal}
