% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-entity-element.R
\name{add_other_entity}
\alias{add_other_entity}
\title{Add Other Entity}
\usage{
add_other_entity(parent_element, other_entity_metadata)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{other_entity_metadata}{A named list or dataframe containing other entity metadata elements
(file_name, file_description, file_type, physical): see \code{\link{create_other_entity}}}
}
\value{
The dataset list or project with other entity file information appended.
}
\description{
Adds the other elements to a dataset list according to EML standards.
}
\examples{
 
other_entity_metadata <- list("file_name" = "other_entity.zip",
                        "file_description" = "An other entity File",
                        )
\dontrun{
dataset <- list() \%>\%
    add_other_entity(other_entity_metadata)
dataset
}
}
