% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-element.R
\name{add_project}
\alias{add_project}
\title{Add Project}
\usage{
add_project(
  parent_element,
  funding_metadata,
  project_title = NULL,
  project_lead = NULL
)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{funding_metadata}{see \code{\link{create_project}}}

\item{project_title}{see \code{\link{create_project}}}

\item{project_lead}{see \code{\link{create_project}}}
}
\description{
Add Project
}
