% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{add_ratio}
\alias{add_ratio}
\title{Add Ratio Measurement Scale}
\usage{
add_ratio(units, unit_precision, number_type, minimum = NULL, maximum = NULL)
}
\arguments{
\item{units}{The units assigned to this attribute.}

\item{unit_precision}{How precise this attirbutes' measurements are recorded.}

\item{number_type}{What type of number. Examples given in exported documentation.}

\item{minimum}{Optional. A theoreical minimum.}

\item{maximum}{Optional. A theoretical maximum.}
}
\description{
Add Ratio Measurement Scale
}
\keyword{internal}
