% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic-coverage-element.R
\name{add_taxonomic_coverage}
\alias{add_taxonomic_coverage}
\title{Add Taxonomic Coverage}
\usage{
add_taxonomic_coverage(
  parent_element,
  CVPIA_common_species = NULL,
  kingdom = "KINGDOM",
  kingdom_value,
  phylum = "PHYLUM",
  phylum_value,
  class = "CLASS",
  class_value,
  order = "ORDER",
  order_value,
  family = "FAMILY",
  family_value,
  genus = "GENUS",
  genus_value,
  species = "SPECIES",
  species_value,
  common_name
)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset}

\item{CVPIA_common_species}{A list of common CVPIA species can be set as the default species so that taxonomic coverage does not need to be inputted in manually. These species include: chinook, steelhead, smelt, and sturgeon.}

\item{kingdom}{Kingdom level present}

\item{kingdom_value}{The kingdom level of the taxonomy.}

\item{phylum}{Phylum level present}

\item{phylum_value}{The phylum level of the taxonomy.}

\item{class}{Class level present}

\item{class_value}{The class level of the taxonomy.}

\item{order}{Order level present}

\item{order_value}{The order level of the taxonomy.}

\item{family}{Family level present}

\item{family_value}{The family level of the taxonomy.}

\item{genus}{Genus level present}

\item{genus_value}{The genus level of the taxonomy.}

\item{species}{Species level present}

\item{species_value}{The species level of the taxonomy.}

\item{common_name}{The common name of the organism}
}
\value{
The dataset or project with taxonomic coverage information appended
}
\description{
Adds the taxonomic coverage information of a dataset based off of EML standards. The addition of taxonomic coverage is optional, however defaults of chinook, sturgeon, smelt, and steelhead are provided. While single or multiple taxonomies can be applied, the full set of information must be provided if chosen to be included.
}
\examples{
add_taxonomic_coverage(parent_element = list(), CVPIA_common_species = "chinook")

add_taxonomic_coverage(parent_element = list(), CVPIA_common_species = "smelt")

add_taxonomic_coverage(parent_element = list(), CVPIA_common_species = "sturgeon")

add_taxonomic_coverage(parent_element = list(), CVPIA_common_species = "steelhead")

add_taxonomic_coverage(parent_element = list(),
                       kingdom = "KINGDOM", kingdom_value = "Animalia",
                       phylum = "PHYLUM", phylum_value = "Chordata",
                       class = "CLASS", class_value = "Mammalia",
                       order = "ORDER", order_value = "Carnivora",
                       family = "FAMILY", family_value = "Felidae",
                       genus = "GENUS", genus_value = "Panthera", 
                       species = "SPECIES", species_value = "Panthera Leo",
                       common_name = "Lion")
}
