% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-table-element.R
\name{attributes_and_codes}
\alias{attributes_and_codes}
\title{Attribute helper function to input into pmap}
\usage{
attributes_and_codes(
  attribute_name,
  attribute_definition,
  storage_type,
  measurement_scale,
  domain,
  type,
  units,
  unit_precision,
  number_type,
  date_time_format,
  date_time_precision,
  minimum,
  maximum,
  attribute_label
)
}
\description{
Attribute helper function to input into pmap
}
\keyword{internal}
