% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-table-element.R
\name{create_data_table}
\alias{create_data_table}
\title{Create Data Table}
\usage{
create_data_table(
  filename,
  attribute_info,
  datatable_description,
  datatable_url = NULL,
  dataset_methods = NULL,
  additional_info = NULL
)
}
\arguments{
\item{filename}{Name of the file you would like added to the datatable element}

\item{attribute_info}{File path to the attribute sheet that contains metadata on attributes for your file}

\item{datatable_description}{Short description of contents of datatable you are adding}

\item{datatable_url}{The URL to the datatable you are adding}

\item{dataset_methods}{Optional if there are methods that are specific to a datatable}

\item{additional_info}{Option if there is additional metadata information that is important to the datatable}

\item{methods}{A set of information of the specific methods used to collect
information in this entity.}
}
\value{
A data table element listed in a formate to be added to a dataset element and turned into EML
}
\description{
Creates the data table elements according to EML standards.
}
\examples{

}
