% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{create_datetime}
\alias{create_datetime}
\title{Create an attribute in dateTime Measurement Scale}
\usage{
create_datetime(date_time_format, date_time_precision, minimum, maximum)
}
\arguments{
\item{date_time_format}{ISO 8601 format should be used.}

\item{date_time_precision}{To what level of time your attribute is recorded.}

\item{minimum}{The earliest dateTime recorded. Values can be larger than
or equal to this number.}

\item{maximum}{The latest dateTime recorded. Values can be less than
or equal to this number.}
}
\description{
Create an attribute in dateTime Measurement Scale
}
\examples{
\dontrun{create_attribute(attribute_name = "Yrs",
                          attribute_definition = "Calendar year of the observation from years 1990 - 2010.",
                          measurement_scale = EMLaide::measurement_scale$dateTime,
                          attribute_label = "Years",
                          date_time_format = "YYYY", 
                          date_time_precision = "1", 
                          minimum = "1993", 
                          maximum = "2003")}
}
\keyword{internal}
