% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding-element.R
\name{create_funding}
\alias{create_funding}
\title{Create Funding Element}
\usage{
create_funding(
  funder_name,
  award_title,
  funder_identifier = NULL,
  award_number = NULL,
  award_url = NULL,
  funding_description = NULL
)
}
\arguments{
\item{funder_name}{Organization or individual providing the funding. If applicable,
use one of the following from the helper data
\code{\link{CVPIA_funders}}: "USBR", "CDWR", or "CDFW" to get pre-selected funding
information. If you are funded by one of these sources
please specify which one in \code{funder_name} and include an \code{award_title}.}

\item{award_title}{Title of the award granted to the project.}

\item{funder_identifier}{(Optional) This is where the funding organization is listed in
the registry. The funder identifier must be registered. Follow the instructions at
\url{https://ror.org/curation/} to get registered.}

\item{award_number}{(Optional) The identifier assigned by the funding agency to identify this funding award.}

\item{award_url}{(Optional) Optional to include a link to information about the funding
award on the funding organization's webpage.}

\item{funding_description}{(Optional) Provide a short description of the funding received.}
}
\value{
An award list that is then added to the project element of an EML file.
}
\description{
Creates the award information of a project based off of EML standards.
This award element is then nested within a project node to complete a funding section.
}
\examples{
create_funding(funder_name = "National Science Foundation",
               funder_identifier = "http://dx.doi.org/10.13039/100000001",
               award_number = "1656026",
               award_title = "National Science Foundation Grant",
               award_url = "https://www.nsf.gov/awardsearch/showAward?AWD_ID=1656026",
               funding_description = "BLE LTER is supported by the National Science 
                                     Foundation under award #1656026 
                                      (2017-08-01 to 2022-07-31)." )
}
