% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword-set-element.R
\name{create_keyword_set}
\alias{create_keyword_set}
\title{Creates Keyword Set Element}
\usage{
create_keyword_set(keyword_metadata)
}
\arguments{
\item{keyword_metadata}{A  dataframe containing \code{keyword} and \code{keywordThesuraus} (if keyword is from a controlled vocabulary).
The dataframe must have a \code{keyword} column and a \code{keywordThesaurus} column. If keywords are not from a controlled vocabulary please leave \code{keywordThesaurus} blank or NA.}
}
\value{
The keyword list that can be appended to a dataset or project list.
}
\description{
Takes in keywords and associated keyword thesaurus and creates a keyword set according to EML standards.
}
\section{Controlled Vocabularies}{


In order to promote consistency, please search the following resources for keywords:

\href{https://vocab.lternet.edu/vocab/vocab/index.php}{LTER} - Long Term Ecological Research.

\href{http://aims.fao.org/standards/agrovoc/functionalities/search}{AGROVOC} -
A controlled vocabulary covering all areas of interest of the Food and Agriculture
Organization (FAO) of the United Nations, including food, nutrition, agriculture,
fisheries, forestry, environment.

\href{https://geonames.usgs.gov/apex/f?p=138:1:5668294677959}{U.S. Board on Geographic Names} -
USGS place names dictionary.
}

\examples{
keyword_metadata <- dplyr::tibble(keyword = c("CVPIA", "dog", "shark", "cat"), 
                                  keywordThesaurus = c(NA, "pet", "ocean", "pet"))
create_keyword_set(keyword_metadata)

keyword_metadata <- dplyr::tibble(keyword = c("CVPIA", "dog", "shark", "cat"), 
                                  keywordThesaurus = c(NA, NA, NA, NA))                                   
create_keyword_set(keyword_metadata)
}
