% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance-element.R
\name{create_maintenance}
\alias{create_maintenance}
\title{Create Maintenance Element}
\usage{
create_maintenance(status = c("complete", "ongoing"), update_frequency = NULL)
}
\arguments{
\item{status}{Provide the status of your project or dataset as either complete or ongoing.}

\item{update_frequency}{Only needed if the status of the project or dataset is ongoing.
If this is the case, please provide the frequency of which the project or dataset is updated.}
}
\value{
A maintenance element that can be appended to a dataset list.
}
\description{
Creates the maintenance information of a dataset based off of EML standards.
}
\details{
Create Maintenance
}
\examples{
create_maintenance(status = "complete")

create_maintenance(status = "ongoing",
                   update_frequency = "Data are updated 
                   annually at the end of the calendar year.")
}
