% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-element.R
\name{create_method}
\alias{create_method}
\title{Create Data Collection Method Element}
\usage{
create_method(
  methods_file,
  instrumentation = character(),
  software = NULL,
  sampling_file = NULL,
  sampling_coverage = NULL,
  sampling_citation = NULL,
  qualityControl_file = NULL
)
}
\arguments{
\item{methods_file}{Please provide a word or markdown document with proper formatting. The
function will then append it properly. An example word document can be found in
"~/EMLaide/inst/extdata/methods-template.docx".}

\item{instrumentation}{Optional. What instruments are being used to conduct the method.}

\item{software}{Optional.}

\item{sampling_file}{Optional.}

\item{sampling_coverage}{Optional.}

\item{sampling_citation}{Optional.}

\item{qualityControl_file}{Optional.}
}
\value{
The methods information.
}
\description{
Creates the data collection method information of a dataset based off of EML standards.
}
\examples{
create_method(methods_file = word_example("methods-template.docx"),
              instrumentation = "Thermometer")
 
}
