% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{create_nominal}
\alias{create_nominal}
\title{Create an Attribute in Nominal Measurement Scale}
\usage{
create_nominal(
  domain = c("text", "enumerated"),
  definition,
  text_pattern = NULL
)
}
\arguments{
\item{domain}{Either "text" or "enumerated".}

\item{definition}{If domain is "text", provide a word description of your attribute.
If the domain is "enumerated", provide a list of the attributes code definitions.}

\item{text_pattern}{Optional. Only needed if applicable and the domain given is "text".}
}
\description{
Create an Attribute in Nominal Measurement Scale
}
\examples{
# Nominal(text):
\dontrun{create_attribute(attribute_name = "site_id",
                          attribute_definition = "Site id as used in sites table",
                          measurement_scale = EMLaide::measurement_scale$nominal,
                          domain= "text",
                          definition = "Site id as used in sites table.")}

# Nominal(enumerated):
code_def_1 = list(code = "yes", definition = "has been captured previously")
code_def_2 = list(code = "no", definition = "has not been captured previously")
code_definition = list(code_def_1, code_def_2)
\dontrun{create_attribute(attribute_name = "Recap", 
                          attribute_definition = "Has the Turtle been captured and tagged previously",
                          measurement_scale = EMLaide::measurement_scale$nominal, 
                          domain = "enumerated",
                          definition = code_definition)}
}
\keyword{internal}
