% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{create_ordinal}
\alias{create_ordinal}
\title{Create an attribute in Ordinal Measurement Scale}
\usage{
create_ordinal(
  domain = c("text", "enumerated"),
  definition,
  text_pattern = NULL
)
}
\arguments{
\item{domain}{Either "text" or "enumerated".}

\item{definition}{If domain is "text", provide a word description of your attribute.
If the domain is "enumerated", provide a list of the attributes code definitions.
Examples are seen in exported documentation.}

\item{text_pattern}{Optional. Only needed if applicable and the domain given is "text".}
}
\description{
Create an attribute in Ordinal Measurement Scale
}
\examples{
# Ordinal(text):
\dontrun{create_attribute(attribute_name = "LatitudeDD", 
                          attribute_definition = "Latitude",
                          measurement_scale = EMLaide::measurement_scale$ordinal,
                          domain= "text", 
                          definition = "Latitude")}
              
# Ordinal(enumerated): 
code_def_0 = list(code = "0", definition = "0 insects per meter of branch")
code_def_1 = list(code = "1", definition = "1-10 insects per meter")
code_def_2 = list(code = "2", definition = "11 – 100 insects per meter")
code_def_3 = list(code = "3", definition = "more than 100 insects per meter")
code_definition = list(code_def_0, code_def_1, code_def_2, code_def_3)
\dontrun{create_attribute(attribute_name = "hwa",
                          attribute_definition = "Hemlock woolly adelgid density per meter of branch",
                          measurement_scale = EMLaide::measurement_scale$ordinal,
                          domain = "enumerated",
                          definition = code_definition)}
}
\keyword{internal}
