% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-entity-element.R
\name{create_other_entity}
\alias{create_other_entity}
\title{Create Other Entity Element}
\usage{
create_other_entity(file_name, file_description, file_type, physical)
}
\arguments{
\item{file_name}{The name of your entity file}

\item{file_description}{A short description of your file}

\item{file_type}{Type of file added as another entity}

\item{physical}{Physical describtion of object use \code{physical} to generate}

\item{parent_element}{A list to append \code{otherEntity} to}
}
\value{
A list that contains all the required elements of the otherEntity
section of an EML document.
}
\description{
This function creates all the required elements for a \code{otherEntity} section of
an ELM document. This should be used if data cannot be added using another more structured format.
}
\details{
Create Other Entity Element
}
