% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnel-element.R
\name{create_person}
\alias{create_person}
\title{Create Personnel Element}
\usage{
create_person(role, first_name, last_name, email, organization, orcid = NULL)
}
\arguments{
\item{role}{Use "creator" if you are one of the primary originators of the data.
Other possible roles "Data Manager", "Field Technician", or "Assistant Researcher".
There can be multiple personnel on a project with the same role.}

\item{first_name}{Person's given name}

\item{last_name}{Person's surname}

\item{email}{Person's email address}

\item{organization}{Person's employer or the entity they are associated
with for this dataset or project}

\item{orcid}{(Optional) ORCID iD is a persistent digital identifier for researchers,
register at \url{http://orcid.org/}}
}
\value{
A person element that can be appended to a dataset or project list.
}
\description{
Create personnel element according to EML standards
}
\examples{
create_person(first_name = 'Katherine', 
              last_name = "Johnson", 
              email = 'kjohnson@nasa.gov', 
              role = 'creator', 
              organization = 'NASA',
              orcid = '12345')

create_person(first_name = "Edith", 
              last_name = "Windsor", 
              email = 'ewindsor@ibm.com', 
              role = 'Data Manager', 
              organization = 'IBM')
}
