% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-date-element.R
\name{create_pub_date}
\alias{create_pub_date}
\title{Create Publication Date}
\usage{
create_pub_date(date = NULL)
}
\arguments{
\item{date}{The publication date. If left null, it will be assigned the current
date. This can be overridden by putting in a specific date.}

\item{parent_element}{A list representing the EML project or dataset.}
}
\value{
A publication date object that can be appended to a dataset or project.
}
\description{
Creates publication date according to EML standards.
}
\examples{
create_pub_date(date = "2020-08-19")
             
# Creates current date:              
create_pub_date() 
}
