% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CVPIA_common_species}
\alias{CVPIA_common_species}
\title{CVPIA Common Species}
\format{
A named list, keys are named the same as values.
}
\usage{
CVPIA_common_species
}
\arguments{
\item{delta_smelt}{}

\item{white_sturgeon}{}
}
\description{
A list of CVPIA common species that can be appended to the taxonomic
coverage.
}
\examples{
CVPIA_common_species$chinook #"chinook"

}
\keyword{datasets}
