% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measurement_scale}
\alias{measurement_scale}
\title{Attribute Measurement Scales}
\format{
A named list, keys are named the same as values.
}
\usage{
measurement_scale
}
\description{
The 5 options for measurement scales which can be appended to the dataset.
A helper data object to be used with \code{\link{add_attribute}}
}
\section{Measurement Scales}{

\itemize
\item{\strong{nominal}}: Used to define categorical scale attributes.
\item{\strong{ordinal}}: Used to define ordered scale attributes.
\item{\strong{interval}}: Used to define interval scale attributes.
\item{\strong{ratio}}: Used to define ratio scale attributes.
\item{\strong{dateTime}}: Used to define date and time attributes.
}

\examples{
measurement_scale$nominal #"nominal"

}
\keyword{datasets}
