% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-functions.R
\name{poll_endpoint_at_dynamic_interval}
\alias{poll_endpoint_at_dynamic_interval}
\title{Poll Endpoint at Dynamic Interval}
\usage{
poll_endpoint_at_dynamic_interval(
  endpoint,
  user_id,
  password,
  init_sleep = 2,
  grow_by = 2,
  verbose = FALSE
)
}
\arguments{
\item{endpoint}{to poll}

\item{user_id}{user id for auth}

\item{password}{password for auth}

\item{init_sleep}{the sleep in seconds on first iteration}

\item{grow_by}{the multiple to grow init_sleep and subsequent sleep amount by}
}
\description{
Poll Endpoint at Dynamic Interval
}
\keyword{internal}
