% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel-element.R
\name{read_excel_sheets}
\alias{read_excel_sheets}
\title{Add Full Excel Document}
\arguments{
\item{excel_path}{The path to the excel document that you wish to read in.}
}
\value{
A list that contains a tibble for each sheet in the excel file.
}
\description{
This function takes in an excel doc with multiple sheets and creates a list
that contains a tibble for each sheet in the excel file.The individual sheets
no longer need to be read in by separate read_excel calls but can instead by accessed
by name from the list generated.
}
\examples{
read_excel_sheets(excel_path = "data-raw/template/template.xlsx")
}
