% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-table-element.R
\name{reorder_attributes_helper}
\alias{reorder_attributes_helper}
\title{Reorder helper function}
\usage{
reorder_attributes_helper(
  attribute_metadata_to_reorder,
  index_of_csv_col_order
)
}
\description{
Reorder helper function
}
\keyword{internal}
