% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-functions.R
\name{reserve_edi_id}
\alias{reserve_edi_id}
\title{Reserve EDI ID}
\usage{
reserve_edi_id(user_id, password)
}
\arguments{
\item{user_id}{A string that contains the user ID for the EDI data portal. If you
do not already have an EDI account you may create one on the EDI portal. \url {https://portal.edirepository.org/nis/login.jsp}}

\item{password}{A string that contains the user password for the EDI data portal.}
}
\value{
This function returns a edi identifier number.
}
\description{
This package returns a funder
}
\examples{
reserve_edi_id(user_id = "samuelwright", 
               password = "340account")
}
