% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{storage_type}
\alias{storage_type}
\title{Attribute Storage Types}
\format{
A named list, keys are named the same as values.
}
\usage{
storage_type
}
\description{
A set of common options for different attribute storage types.
A helper data object to be used with \code{\link{add_attribute}}
}
\section{Character Storage Types}{


\itemize
\item{\strong{string}}: The string datatype represents character strings in XML.
Use \href{https://www.w3.org/TR/xmlschema-2/#string}{this} link for more information.
}

\section{Locical Storage Types}{


\itemize
\item{\strong{boolean}}: The boolean datatype has the value space required
to support the mathematical concept of binary-valued logic i.e {true, false}.
Use \href{https://www.w3.org/TR/xmlschema-2/#boolean}{this} link for more information.
}

\section{Numeric Storage Types}{


\itemize
\item{\strong{decimal}}: The decimal datatype represents a subset of the real numbers,
which can be represented by decimal numerals. Use
\href{https://www.w3.org/TR/xmlschema-2/#decimal}{this} link for more information.
\item{\strong{float}}: The float datatype is patterned after the IEEE
single-precision 32-bit floating point type. Use \href{https://www.w3.org/TR/xmlschema-2/#float}{this}
link for more information.
\item{\strong{double}}: The double datatype is patterned after the IEEE
double-precision 64-bit floating point type. Use \href{https://www.w3.org/TR/xmlschema-2/#double}{this}
link for more information.
}

\section{Time Storage Types}{


\itemize
\item{\strong{duration}}:The duration datatype represents a duration of time.
The value space of duration is a six-dimensional space where the coordinates
designate the Gregorian year, month, day, hour, minute, and second components
defined in ISO 8601. Use \href{https://www.w3.org/TR/xmlschema-2/#duration}{this} link for more information.
\item{\strong{dateTime}}:The dateTime datatype values may be viewed as objects
with integer-valued year, month, day, hour and minute properties, a
decimal-valued second property, and a boolean timezoned property. Use
\href{https://www.w3.org/TR/xmlschema-2/#dateTime}{this} link for more information.
\item{\strong{time}}: The time datatype represents an instant of time that recurs every day.
Use \href{https://www.w3.org/TR/xmlschema-2/#time}{this} link for more information.
\item{\strong{date}}: The date datatype consists of top-open intervals of
exactly one day in length on the timelines of dateTime, beginning on the
beginning moment of each day (in each timezone), i.e. '00:00:00', up to but not
including '24:00:00'. Use \href{https://www.w3.org/TR/xmlschema-2/#date}{this}
link for more information.
\item{\strong{gYearMonth}}: The gYearMonth datatype represents a specific
gregorian month in a specific gregorian year. Use \href{https://www.w3.org/TR/xmlschema-2/#gYearMonth}{this}
link for more information.
\item{\strong{gYear}}: The gYear datatype represents a gregorian calendar year.
Use \href{https://www.w3.org/TR/xmlschema-2/#gYear}{this} link for more information.
\item{\strong{gMonthDay}}: The gMonthDay datatype is a gregorian date that
recurs, specifically a day of the year such as the third of May.
Use \href{https://www.w3.org/TR/xmlschema-2/#gMonthDay}{this} link for more information.
\item{\strong{gDay}}: The gDay datatype is a gregorian day that recurs,
specifically a day of the month such as the 5th of the month. Use
\href{https://www.w3.org/TR/xmlschema-2/#gDay}{this} link for more information.
\item{\strong{gMonth}} The gMonth datatype is a gregorian month that recurs every year.
Use \href{https://www.w3.org/TR/xmlschema-2/#gMonth}{this} link for more information.
}

\examples{
storage_type$integer #"integer"
}
\keyword{datasets}
