% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-functions.R
\name{update_edi_package}
\alias{update_edi_package}
\title{Update EDI Data Package}
\usage{
update_edi_package(
  user_id,
  password,
  existing_package_identifier,
  eml_file_path
)
}
\arguments{
\item{user_id}{EDI data portal user ID. Create an account an
EDI \href{https://portal.edirepository.org/nis/login.jsp}{here}}

\item{password}{EDI data portal user password}

\item{existing_package_identifier}{The current edi number of the package that you are trying to update.
Do not include the "edi" prefix, just put the number. (ex: For "edi.101.1" put 101.1)}

\item{eml_file_path}{The file path to the EML metadata document that you wish to use to update.
(A web link to the csv must be included in the dataset information in the EML in order for a data package to be evaluated.)}
}
\value{
Message describing if your package was successfully updated or not.
}
\description{
This function takes in authentication info for EDI, a package number, and an updated EML file to
updated an existing package on EDI.
}
\examples{
\dontrun{update_edi_package(user_id = "samuelwright", 
                            existing_package_identifier = "740.1",
                            eml_file_path = "data/edi20.1.xml")}
}
