% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-functions.R
\name{upload_edi_package}
\alias{upload_edi_package}
\title{Upload EDI Data Package}
\usage{
upload_edi_package(
  user_id,
  password,
  eml_file_path,
  environment = "production",
  package_size = "small"
)
}
\arguments{
\item{user_id}{EDI data portal user ID. Create an account an
EDI \href{https://portal.edirepository.org/nis/login.jsp}{here}}

\item{password}{EDI data portal user password}

\item{eml_file_path}{The file path to the EML metadata document that you wish to evaluate.
(A web link to the csv must be included in the dataset information in the EML in order for a data package to be evaluated.)}

\item{environment}{EDI portal environment to run command in. Can be: "production" - environment for publishing to EDI ,
"staging" - environment to test upload and rendering of new environment, "development"}
}
\value{
Message describing if your package was successfully updated or not.
}
\description{
This function takes in authentication info for EDI and an EML file to
be evaluated uploaded to EDI.
}
\details{
For more information about the validation services see \href{https://pastaplus-core.readthedocs.io/en/latest/doc_tree/pasta_api/data_package_manager_api.html#upload-and-evaluation}{the PASTAplus docs}
}
\examples{
\dontrun{upload_edi_package(user_id = "samuelwright", 
                            eml_file_path = "data/edi20.1.xml")}
}
