% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-example-helper.R
\name{word_example}
\alias{word_example}
\title{Word Document Example Reader}
\usage{
word_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
Helps read documents which are present in the examples of the
\code{cvpiaEDIutils} package.
}
\examples{
word_example()
word_example("abstract_template.docx")
}
