% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-element.R
\name{add_abstract}
\alias{add_abstract}
\title{Add Abstract Element}
\usage{
add_abstract(parent_element, abstract)
}
\arguments{
\item{abstract}{Abstract paragraphs, requires a minimum of 20 words. A template
word document is accessible in "~/EDIutils/inst/extdata/abstract_template.docx".
You can copy and paste your abstract in this document and set the parameter of
\code{abstract} equal to this file path.}

\item{parament_element}{A list representing the EML project or dataset.}
}
\value{
The dataset or project with abstract appended
}
\description{
Adds the abstract of a dataset according to EML standards.
}
\details{
For	a	dataset,	the	abstract	element	can	appear	at	the	resource	level	or	the	project level.
The	\code{abstract}	element	will	be	used	for	full-text	searches,	and	it	should	be	rich	with
descriptive	text.	In	particular,	descriptions	should	include	information	that	does	not	fit	into
structured	metadata,	and	focus	on	the	\code{"what"},	\code{"when"},	and	\code{"where"}	information,	general
taxonomic	information,	as	well	as	whether	the	dataset	is	ongoing	or	completed.	Some
general	methods	description	is	appropriate,	and	broad	classes	of	measured	parameters
should	also	be	included.		For	a	large	number	of	parameters,	use	categories	instead	of	listing
all	parameters	(e.g.	use	the	term	\code{"nutrients"}	instead	of	nitrate,	phosphate,	calcium,	etc.),	in
combination	with	the	parameters	that	seem	most	relevant	for	searches.
}
\examples{
add_abstract(parent_element = list(),
             abstract = word_example("abstract_template.docx"))
          
}
