% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{add_attribute}
\alias{add_attribute}
\title{Add Atribute}
\usage{
add_attribute(
  attribute_name,
  attribute_definition,
  storage_type,
  measurement_scale,
  attribute_label = NULL,
  domain = NULL,
  definition = NULL,
  text_pattern = NULL,
  units = NULL,
  number_type = NULL,
  unit_precision = NULL,
  date_time_format = NULL,
  date_time_precision = NULL,
  minimum = NULL,
  maximum = NULL
)
}
\arguments{
\item{attribute_name}{The name    of a field in    a    data table. This is    often    a short
and/or cryptic name. It is recommended that the attribute names    be suitable    for
use    as a    variable,    e.g.,    composed of ASCII    characters,    and    that the attribute
names match the column headers of a CSV or other text table.}

\item{storage_type}{An indicator    to processing    systems    as to    how    the    attribute
might be represented in    a    system or    language,    but    is distinct    from the actual
expression of    the    domain of    the    attribute. A list of approved storage types can be found at
\code{\link{storage_type}}}

\item{measurement_scale}{The    type of    scale    from which values are    drawn    for    the
attribute. A list of approved measurment scales can be viewed at \code{\link{measurement_scale}}.
Further explination of the choices and required inputs are described below.}

\item{attribute_label}{Optional. Used to provide    a    less ambiguous or    less cryptic
alternative identification    than what    is provided    in \code{attribute_name}.}

\item{domain}{Input for the non-numeric measurement scales only.
Please list either "text" or "enumerated".}

\item{definition}{Either the text definition or code definition. Both are appended
using this parameter, but have different applications. Examples of both are listed below.}

\item{text_pattern}{A regular expression pattern constraining the attribute.}

\item{units}{The units assigned to this attribute's values.}

\item{number_type}{A list of possible options can be viewed at \code{\link{storage_type}}.}

\item{unit_precision}{How precise units are measured.}

\item{date_time_format}{The format your date/time attribute is recorded in.
ISO 8601 standard should be used (YYYY-MM-DD).}

\item{date_time_precision}{To what level time is being measured.}

\item{minimum}{Theoretical or allowable minimum value.}

\item{maximum}{Theoretical or allowable maximum value.}

\item{attribute_defenition}{A precise and complete    definition of the attribute    being    documented.}
}
\value{
The project or dataset list with an attribute list appended
}
\description{
Adds an attribute element according to EML standards. Multiple
attributes can be created. Will be appended to the attribute list.
}
\section{Measurement Scales}{

Different measurement scale values will indicate different inputs:

\subsection{Non-numeric:} Please provide a domain to indicate if your attribute is text or enumerated.
\emph{Nominal:}
Used to define categorical scale attributes. If your attribute falls
under the domain of "text", please provide the inputs of \code{definition} and \code{text_pattern}.
If your attribute falls under the domain of "enumerated", please provide the input of \code{definition}.

\emph{Ordinal:}
Used to define ordered scale attributes. If your attribute falls under the domain
of "text", please provide the inputs of \code{definition} and \code{text_pattern}. If your
attribute falls under the domain of "enumerated", please provide the input of \code{definition}.

\subsection{Numeric:}

\emph{Interval:}
Used to define interval scale attributes. Please provide the inputs of \code{units},
\code{unit_precision}, \code{number_type}, \code{minimum}, and \code{maximum}.

\emph{Ratio:}
Used to define ratio scale attributes. Please provide the inputs of \code{units},
\code{unit_precision}, \code{number_type}, \code{minimum}, and \code{maximum}.

\emph{dateTime:}
Used to define date and time attributes. Please provide the inputs of \code{date_time_format},
\code{date_time_precision}, \code{minimum}, and \code{maximum}.
}

