% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{add_attribute}
\alias{add_attribute}
\title{Add Atribute}
\usage{
add_attribute(
  attribute_name,
  attribute_definition,
  storage_type,
  measurement_scale,
  attribute_label = NULL,
  domain = NULL,
  definition = NULL,
  text_pattern = NULL,
  type = NULL,
  units = NULL,
  number_type = NULL,
  unit_precision = NULL,
  date_time_format = NULL,
  date_time_precision = NULL,
  minimum = NULL,
  maximum = NULL
)
}
\arguments{
\item{attribute_name}{The name    of a field in    a    data table. This is    often    a short
and/or cryptic name. It is recommended that the attribute names    be suitable    for
use    as a    variable,    e.g.,    composed of ASCII    characters,    and    that the attribute
names match the column headers of a CSV or other text table.}

\item{storage_type}{An indicator    to processing    systems    as to    how    the    attribute
might be represented in    a    system or    language,    but    is distinct    from the actual
expression of    the    domain of    the    attribute. A list of approved storage types can be found at
\code{\link{storage_type}}}

\item{measurement_scale}{The    type of    scale    from which values are    drawn    for    the
attribute. A list of approved measurment scales can be viewed at \code{\link{measurement_scale}}.
Further explination of the choices and required inputs are described below.}

\item{attribute_label}{(Optional) Used to provide    a    less ambiguous or    less cryptic
alternative identification    than what    is provided    in \code{attribute_name}.}

\item{domain}{Input for the non-numeric measurement scales only.
Please list either "text" or "enumerated".}

\item{definition}{Either the text definition or code definition. Both are appended
using this parameter, but have different applications. Examples of both are listed below.}

\item{text_pattern}{(Optional) A regular expression pattern constraining the attribute.}

\item{type}{Either "ratio" or "interval".}

\item{units}{The units assigned to this attribute's values.}

\item{number_type}{A list of possible options can be viewed at \code{\link{number_type}}.}

\item{unit_precision}{(Optional) How precise units are measured.}

\item{date_time_format}{The format your date/time attribute is recorded in.
ISO 8601 standard should be used (YYYY-MM-DD).}

\item{date_time_precision}{To what level time is being measured.}

\item{minimum}{Theoretical or allowable minimum value. Values can be larger than
or equal to this number.}

\item{maximum}{Theoretical or allowable maximum value. Values can be less than
or equal to this number.}

\item{attribute_defenition}{A precise and complete    definition of the attribute    being    documented.}
}
\value{
The project or dataset list with an attribute list appended.
}
\description{
Adds an attribute element according to EML standards. Multiple
attributes can be created. Will be appended to the attribute list.
}
\section{Measurement Scales}{

Different measurement scale values will indicate different inputs:
\subsection{Non-numeric:}{

Please provide a domain to indicate if your attribute is text or enumerated.

\emph{Nominal:}
Used to define categorical scale attributes. If your attribute falls
under the domain of "text", please provide the inputs of \code{definition} and \code{text_pattern}.
If your attribute falls under the domain of "enumerated", please provide the input of \code{definition}.

\emph{Ordinal:}
Used to define ordered scale attributes. If your attribute falls under the domain
of "text", please provide the inputs of \code{definition} and \code{text_pattern}. If your
attribute falls under the domain of "enumerated", please provide the input of \code{definition}.
}

\subsection{Numeric:}{

\emph{Interval:}
Used to define interval scale attributes. Please provide the inputs of \code{type},
\code{units}, \code{unit_precision}, \code{number_type}, \code{minimum}, and \code{maximum}.

\emph{Ratio:}
Used to define ratio scale attributes. Please provide the inputs of \code{type},
\code{units}, \code{unit_precision}, \code{number_type}, \code{minimum}, and \code{maximum}.

\emph{dateTime:}
Used to define date and time attributes. Please provide the inputs of \code{date_time_format},
\code{date_time_precision}, \code{minimum}, and \code{maximum}.
}
}

\examples{
# Nominal(text):
add_attribute(attribute_name = "site_id",
              attribute_definition = "Site id as used in sites table",
              storage_type = EDIutils::storage_type$integer,
              measurement_scale = EDIutils::measurement_scale$nominal,
              domain= "text",
              definition = "Site id as used in sites table.")

# Nominal(enumerated):
code_def_1 = list(code = "yes", definition = "has been captured previously")
code_def_2 = list(code = "no", definition = "has not been captured previously")
code_definition = list(code_def_1, code_def_2)
add_attribute(attribute_name = "Recap", 
              attribute_definition = "Has the Turtle been captured and tagged previously",
              storage_type = EDIutils::storage_type$text, 
              measurement_scale = EDIutils::measurement_scale$nominal, 
              domain = "enumerated",
              definition = code_definition)
              
# Ordinal(text):
add_attribute(attribute_name = "LatitudeDD", 
              attribute_definition = "Latitude",
              storage_type = EDIutils::storage_type$string,
              measurement_scale = EDIutils::measurement_scale$ordinal,
              domain= "text", 
              definition = "Latitude")
              
# Ordinal(enumerated): 
code_def_0 = list(code = "0", definition = "0 insects per meter of branch")
code_def_1 = list(code = "1", definition = "1-10 insects per meter")
code_def_2 = list(code = "2", definition = "11 – 100 insects per meter")
code_def_3 = list(code = "3", definition = "more than 100 insects per meter")
code_definition = list(code_def_0, code_def_1, code_def_2, code_def_3)
add_attribute(attribute_name = "hwa",
              attribute_definition = "Hemlock woolly adelgid density per meter of branch",
              storage_type = EDIutils::storage_type$decimal,
              measurement_scale = EDIutils::measurement_scale$ordinal,
              domain = "enumerated",
              definition = code_definition)
              
# Interval:
add_attribute(attribute_name = "Count",
              attribute_definition = "Number of individuals observed",
              measurement_scale = EDIutils::measurement_scale$interval, 
              storage_type = EDIutils::storage_type$integer,
              type = "interval",
              units = "number",
              unit_precision = "1",
              number_type = EDIutils::number_type$whole, 
              minimum = "0")
                   
# Ratio: 
add_attribute(attribute_name = "pH",
              attribute_definition = "pH of soil solution",
              storage_type = EDIutils::storage_type$float,
              measurement_scale = EDIutils::measurement_scale$ratio,
              type = "ratio",
              units = "dimensionless",
              unit_precision = "0.01",
              number_type = EDIutils::number_type$real)
                   
# dateTime:
add_attribute(attribute_name = "Yrs", 
              attribute_definition = "Calendar year of the observation from years 1990 - 2010.",
              storage_type = EDIutils::storage_type$integer,
              measurement_scale = EDIutils::measurement_scale$dateTime,
              attribute_label = "Years",
              date_time_format = "YYYY", 
              date_time_precision = "1", 
              minimum = "1993", 
              maximum = "2003")
}
