% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-list-element.R
\name{add_attribute_list}
\alias{add_attribute_list}
\title{Add Atribute List}
\usage{
add_attribute_list(
  attribute_name,
  attribute_label,
  attribute_definition,
  storage_type,
  measurement_scale,
  nominal_scale_definition = NULL,
  units = NULL,
  number_type = NULL,
  code_number = NULL,
  code_number_definition = NULL,
  unit_precision = NULL,
  ISO_date_time = NULL,
  date_time_precision = NULL,
  minimum = NULL,
  maximum = NULL
)
}
\arguments{
\item{attribute_name}{The name    of a field in    a    data table. This is    often    a short    and/or cryptic name. It is recommended that the attribute names    be suitable    for    use    as a    variable,    e.g.,    composed of ASCII    characters,    and    that the    attribute names    match    the    column headers of    a    CSV    or other text table.}

\item{attribute_label}{Optional. Used to provide    a    less ambiguous or    less cryptic alternative identification    than what    is provided    in attribute_name}

\item{storage_type}{An indicator    to processing    systems    as to    how    the    attribute    might    be represented in    a    system or    language,    but    is distinct    from the actual    expression of    the    domain of    the    attribute. Non system-specific examples    include: float,    integer    and    string.}

\item{measurement_scale}{The    type of    scale    from which values are    drawn    for    the    attribute. Must use one of the following types: nominal, ordinal, interval, ratio, or dateTime.}

\item{nominal_scale_definition}{Description for nominal measurement scale values.}

\item{units}{The units assigned to this particular attributes values.}

\item{number_type}{Can be defined as: real, natural, whole, or integer}

\item{code_number}{The reference number to the code definition.}

\item{code_number_definition}{The explaination of a listed code number}

\item{unit_precision}{How precise units are measured}

\item{ISO_date_time}{YYYY-MM-DDThh:mm:ssTZD format}

\item{date_time_precision}{What level of time is being measured.}

\item{minimum}{Theoretical or allowable minimum value.}

\item{maximum}{Theoretical or allowable maximum value.}

\item{attribute_defenition}{A precise and complete    definition of the attribute    being    documented.}
}
\value{
The project or dataset list with an attribute list appended
}
\description{
adds the attribute list element according to EML standards. Multiple attributes can be created.
}
\examples{
TODO
Different measurement scale values will indicate different inputs: 
   Nominal: Please provide measurement_scale_definition
   Ordinal: Please provide code_number and code_number_definition 
   Interval: Please provide units, unit_precision, number_type, minimum, and maximum
   Ratio: Please provide units and number_type
   dateTime: Please provide ISO_date_time, date_time_precision, minimum, and maximum
}
