% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-list-element.R
\name{add_attribute_list}
\alias{add_attribute_list}
\title{Add Atribute List}
\usage{
add_attribute_list(
  attribute_name,
  attribute_label,
  attribute_definition,
  storage_type,
  measurement_scale,
  measurement_scale_description,
  units,
  number_type
)
}
\arguments{
\item{attribute_name}{TODO}

\item{attribute_label}{TODO}

\item{storage_type}{TODO}

\item{measurement_scale}{TODO}

\item{measurement_scale_description}{TODO}

\item{units}{TODO}

\item{number_type}{TODO}

\item{attribute_defenition}{TODO}
}
\value{
the project or dataset list with an attribute list appended
}
\description{
adds the attribute list element according to EML standards
}
\examples{
TODO
}
