% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage-element.R
\name{add_coverage}
\alias{add_coverage}
\title{Add Coverage}
\usage{
add_coverage(parent_element, coverage_metadata, taxonomic_metadata = NULL)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{coverage_metadata}{a list or datatable of coverage metadata see \code{\link{create_coverage}}}

\item{taxonomic_metadata}{a list or datatable of taxonomic metadata see \code{\link{create_taxonomic_coverage}}}
}
\description{
Add Coverage
}
\examples{
coverage <- list(geographic_description = "Description",
             west_bounding_coordinate = "-160.594000", 
             east_bounding_coordinate = "-134.104800",
             north_bounding_coordinate = "71.238300", 
             south_bounding_coordinate = "67.865000",
             begin_date = "1980-01-01", 
             end_date = "2010-12-31")
             
dataset <- list() \%>\% 
  add_coverage(coverage)
}
