% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-table-element.R
\name{add_data_table}
\alias{add_data_table}
\title{Add Data Table}
\usage{
add_data_table(parent_element, datatable_metadata)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{datatable_metadata}{A named list or dataframe containing datatable elements: see \code{\link{create_data_table}}}
}
\description{
Add Data Table
}
