% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-table-element .R
\name{add_data_table}
\alias{add_data_table}
\title{Add Data Table}
\usage{
add_data_table(
  parent_element,
  alternate_identifier = NULL,
  entity_name,
  entity_description,
  object_name,
  number_of_headers,
  record_delimiter,
  physcial_line_delimiter,
  attribute_orientation,
  field_delimiter,
  online_url,
  attribute_list,
  case_sensitivity,
  number_of_records,
  constraint
)
}
\arguments{
\item{entity_name}{The name of the table, file, or database table. Often thought of as the original ascii file name.}

\item{entity_description}{A longer, more descriptive explination of the data in the entity.}

\item{object_name}{The name of the file when downloaded or exported as text from a database.}

\item{number_of_headers}{TODO}

\item{record_delimiter}{TODO}

\item{attribute_orientation}{TODO}

\item{field_delimiter}{TODO}

\item{online_url}{A link to access the data.}

\item{attribute_list}{Describes all variables in a data entity in individual attribute elements. These descriptions include the name and definition of each attribute, its domain, definition of coded values, and other pertinent information. It is further explained in the add_attribute_list function.}

\item{case_sensitivity}{Designates the data table as case sensitive or not. Please provide "yes" or "no".}

\item{number_of_records}{A count of the number of records in the data table.}

\item{constraint}{TODO. Describes any integrity constraints between entities (e.g.tables), as they would    be maintained    in a relational    management system.}

\item{parament_element}{a list representing the EML project or dataset}

\item{alternate_identifer}{Optional. Provide when a dataset or project belongs to more than the contributing organization. Please include each additional individual ID with its own alternate_identifier.}

\item{physical_line_delimiter}{TODO}
}
\value{
the project or dataset list with a data table appended
}
\description{
adds the data table elements according to EML standards
}
\examples{
TODO
}
