% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{add_datetime}
\alias{add_datetime}
\title{Add dateTime Measurement Scale}
\usage{
add_datetime(date_time_format, date_time_precision, minimum, maximum)
}
\arguments{
\item{date_time_format}{ISO 8601 format should be used.}

\item{date_time_precision}{To what level of time your attribute is recorded.}

\item{minimum}{The earliest dateTime recorded.}

\item{maximum}{The latest dateTime recorded.}
}
\description{
Add dateTime Measurement Scale
}
\examples{
add_attribute(attribute_name = "Yrs",
              attribute_definition = "Calendar year of the observation from years 1990 - 2010.",
              storage_type = cvpiaEDIutils::storage_type$integer,
              measurement_scale = cvpiaEDIutils::measurement_scale$dateTime,
              attribute_label = "Years",
              date_time_format = "YYYY", 
              date_time_precision = "1", 
              minimum = "1993", 
              maximum = "2003")
}
\keyword{internal}
