% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding-element.R
\name{add_funding}
\alias{add_funding}
\title{Adds Funding Element}
\usage{
add_funding(
  parent_element,
  funder_name,
  funder_identifier,
  award_number,
  award_title,
  award_url,
  funding_description
)
}
\value{

}
\description{
Adds Funding Element
}
