% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding-element.R
\name{add_funding}
\alias{add_funding}
\title{Add Funding Element}
\usage{
add_funding(
  CVPIA_default_funder = NULL,
  funder_name,
  funder_identifier = NULL,
  award_number = NULL,
  award_title,
  award_url = NULL,
  funding_description = NULL
)
}
\arguments{
\item{CVPIA_default_funder}{If funding is one of the common CVPIA funding sources select
a default funding source from the \code{CVPIA_default_funder} options in the metadata.xlsx
sheet. The options for default funding are USBR, CDWR, and CDFW.}

\item{funder_name}{Organization or individual providing the funding.}

\item{funder_identifier}{(Optional) This is where the funding organization is listed in
the registry. The funder identifier must be registered. Follow the instructions at
\url{https://ror.org/curation/} to get registered.}

\item{award_number}{(Optional) The identifier assigned by the funding agency to identify this funding award.}

\item{award_title}{Title of the award granted to the project.}

\item{award_url}{(Optional) Optional to include a link to information about the funding
award on the funding organization's webpage.}

\item{funding_description}{(Optional) Provide a short description of the funding received.}
}
\value{
An award list that is then added to the project element of an EML file.
}
\description{
Creates the award information of a project based off of EML standards.
This award element is then nested within a project node to complete a funding section.
}
\examples{
add_funding(funder_name = "National Science Foundation",
            funder_identifier = "http://dx.doi.org/10.13039/100000001",
            award_number = "1656026",
            award_title = "National Science Foundation Grant",
            award_url = "https://www.nsf.gov/awardsearch/showAward?AWD_ID=1656026",
            funding_description = "BLE LTER is supported by the National Science 
                                   Foundation under award #1656026 
                                   (2017-08-01 to 2022-07-31)." )
}
