% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-element.R
\name{add_interval_ratio}
\alias{add_interval_ratio}
\title{Add Interval or Ratio Measurement Scales}
\usage{
add_interval_ratio(
  type = c("interval", "ratio"),
  units,
  unit_precision,
  number_type,
  minimum = NULL,
  maximum = NULL
)
}
\arguments{
\item{type}{Either "interval" or "ratio". Use "interval" to define
data which consist of equidistant points on a scale. Use "ratio" to define data
which consists not only of equidistant points but also has a meaningful zero
point, which allows the ratio to have meaning.}

\item{units}{The units assigned to this attribute.}

\item{unit_precision}{How precise this attirbutes' measurements are recorded.}

\item{number_type}{What type of number. Examples given in exported documentation.}

\item{minimum}{Optional. A theoreical minimum.}

\item{maximum}{Optional. A theoretical maximum.}
}
\description{
Add Interval or Ratio Measurement Scales
}
\examples{
Interval:
add_attribute(attribute_name = "Count",
              attribute_definition = "Number of individuals observed",
                   measurement_scale = cvpiaEDIutils::measurement_scale$interval, 
                   storage_type = cvpiaEDIutils::storage_type$integer,
                   type = "interval"
                   units = "number",
                   unit_precision = "1",
                   number_type = "whole", 
                   minimum = "0")
                   
Ratio: 
add_attribute(attribute_name = "pH",
              attribute_definition = "pH of soil solution",
              storage_type = cvpiaEDIutils::storage_type$float,
              measurement_scale = cvpiaEDIutils::measurement_scale$ratio,
              type = "ratio",
              units = "dimensionless",
              unit_precision = "0.01",
              number_type = "real")
}
\keyword{internal}
