% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword-set-element.R
\name{add_keyword_set}
\alias{add_keyword_set}
\title{Add Keywords}
\usage{
add_keyword_set(parent_element, keyword_metadata)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{keyword_metadata}{A named list or dataframe containing keyword elements: see \code{\link{create_keywords}}}
}
\value{
The dataset list or project with keyword information appended.
}
\description{
Adds the keyword metadata elements to a dataset list according to EML standards.
}
