% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance-element.R
\name{add_maintenance}
\alias{add_maintenance}
\title{Add Maintenance}
\usage{
add_maintenance(parent_element, maintenance_metadata)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{maintenance_metadata}{A table or list containing maintenance metadata: see \code{\link{create_maintenance}}}
}
\value{
The dataset list or project with maintenance information appended.
}
\description{
Adds the maintenance metadata element to a dataset list according to EML standards.
}
\examples{
maintenance_metadata <- list(status = "ongoing", update_frequency = "monthly")
dataset <- list() \%>\%
     add_maintenance(maintenance_metadata) 
dataset
}
