% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance-element.R
\name{add_maintenance}
\alias{add_maintenance}
\title{Add Maintenance Element}
\usage{
add_maintenance(parent_element, status)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{status}{Provide the status of your project or dataset as either complete or ongoing.}
}
\value{
The dataset or project with maintenance information appended.
}
\description{
Adds the maintenance information of a dataset based off of EML standards.
}
\examples{
add_maintenance(parent_element = list(), status = ongoing)
}
