% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-element.R
\name{add_method}
\alias{add_method}
\title{Add Methods}
\usage{
add_method(parent_element, methods_file)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{methods_file}{A file containing methods information (word or markdown): see \code{\link{create_method}}}
}
\value{
The dataset list or project with methods information appended.
}
\description{
Adds the methods metadata elements according to EML standards.
}
\examples{
methods_file = word_example("methods-template.docx")

dataset <- list() \%>\%
    add_method(methods_file)
dataset
}
