% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-element.R
\name{add_method}
\alias{add_method}
\title{Add Data Collection Method Element}
\usage{
add_method(methods_file, instrumentation = character())
}
\arguments{
\item{instrumentation}{Optional. What is being used to conduct the method.}

\item{method_file}{Please provide a word document with proper formatting. The
function will then append it properly. An example word document can be found in
"~/cvpiaEDIutils/data-raw/template/methods-template.docx". This file can be opened
and edited to include your proper methods section.}
}
\value{
The methods information.
}
\description{
Adds the data collection method information of a dataset based off of EML standards.
}
\examples{
add_method(method_file = "~/cvpiaEDIutils/data-raw/template/methods-template.docx",
           instrumentation = "Thermometer")
 
}
