% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-element.R
\name{add_method}
\alias{add_method}
\title{Add Data Collection Method Element}
\usage{
add_method(description, title = NULL, instrumentation = NULL)
}
\arguments{
\item{description}{How this method is being conducted.}

\item{title}{Optional. What the particular method is accomplishing.}

\item{instrumentation}{Optional. What is being used to conduct the method.}
}
\value{
The dataset or project with methods information appended.
}
\description{
Adds the data collection method information of a dataset based off of EML standards.
}
\examples{
add_method(title = "Climate Data",
           description = "Daily temperature (maximum/minimum) and
           precipitation data were obtained for each stand from 1996 to 2011
           from the online PRISM Gridded Climate database (PRISM Climate Group,
           Oregon State University, http://prism.oregonstate.edu, created 26 Mar 2015)
           by interpolating 4km2 resolution climate data at the centroid of each
           eastern hemlock stand using values from surrounding grid cell centers
           and inverse-distance squared weighting.", instrumentation = "Thermometer")
                
If the method has a longer description than one paragraph:
description1 <- "This is the first paragraph" 
description2 <- "This is the second paragraph." 
description = list(description1 description2)
add_method(title = "The Data", description = description,
                instrumentation = "The applicable instrument")
}
