% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-element.R
\name{add_method}
\alias{add_method}
\title{Add Data Collection Method Element}
\usage{
add_method(
  parent_element,
  methods_file,
  instrumentation = character(),
  software = NULL,
  sampling_file = NULL,
  sampling_coverage = NULL,
  sampling_citation = NULL,
  qualityControl_file = NULL
)
}
\arguments{
\item{instrumentation}{Optional. What is being used to conduct the method.}

\item{software}{Optional.}

\item{sampling_file}{Optional.}

\item{sampling_coverage}{Optional.}

\item{sampling_citation}{Optional.}

\item{qualityControl_file}{Optional.}

\item{method_file}{Please provide a word document with proper formatting. The
function will then append it properly. An example word document can be found in
"~/EDIutils/inst/extdata/methods-template.docx". This file can be opened
and edited to include your proper methods section.}
}
\value{
The methods information.
}
\description{
Adds the data collection method information of a dataset based off of EML standards.
}
\examples{
add_method(parent_element = list(), 
           methods_file = word_example("methods-template.docx"),
           instrumentation = "Thermometer")
 
}
