% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-element.R
\name{add_method_step}
\alias{add_method_step}
\title{TODO}
\usage{
add_method_step(method_step_title, method_step_description, instrumentation)
}
\arguments{
\item{method_step_title}{TODO}

\item{method_step_description}{TODO}

\item{instrumentation}{TODO}
}
\description{
TODO
}
\examples{
TODO
}
