% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnel-element.R
\name{add_personnel}
\alias{add_personnel}
\title{Add personnel}
\usage{
add_personnel(parent_element, personnel_metadata)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset}

\item{personnel_metadata}{A dataframe of personnel information see \code{\link{create_person}}}
}
\value{
The dataset list or project with personnel information appended.
}
\description{
Adds the personnel metadata elements to a dataset list according to EML standards.
}
\details{
One person within \code{personnel_metadata} must have role of 'creator'.
}
\examples{
personnel_metadata <- dplyr::tibble(first_name = "Stacy", last_name = "Banet", 
                                    email = "Stacy@aol.com", 
                                    role = "creator", organization = "USBR", orcid = NA)
dataset <- list() \%>\%
   add_personnel(personnel_metadata)
dataset
}
