% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical-element.R
\name{add_physical}
\alias{add_physical}
\title{Add Physical Element}
\usage{
add_physical(
  file_path,
  number_of_headers = "1",
  record_delimiter = "\\\\r\\\\n",
  attribute_orientation = "column",
  field_delimiter = c(",", " ", "\\t", ":"),
  authentication = NULL
)
}
\arguments{
\item{file_path}{The full name of the file of which your data is being read.
From this, the \code{object_name} and \code{object_size} will be calculated.}

\item{number_of_headers}{Number of header lines preceeding the data. A default
of "1" is assigned if no input given as it is bad practice to provide anything
greater than 1.}

\item{record_delimiter}{Character used to delimit records. If no value is
inputted the default value of '\\r\\n' is assigned.}

\item{attribute_orientation}{The orientation of the attributes. A default of
"column" will be assigned if no input is given.}

\item{field_delimiter}{Character used to delimit each field. The options that
can be used are commas (","), space (" "), tab ("\t"), or colon (":").}

\item{authentication}{A value, typically a checksum, used to authenticate that
the bitstream delivered to the user is identical to the original. This also
descirbes authentication procedures or techniques. Can be used to describe
authentication procedures or techniques, typically by giving a checksum value
for the object.}

\item{number_of_footers}{Number of footer lines succeeding the data.}
}
\value{
A complete set of information on the physical format.
}
\description{
Adds the information of the physical format of the dataset based off of EML standards.
}
\examples{
add_physical(file_path = "/Users/lizzyshaw/FlowWest/cvpiaEDIutils/data-raw/Lowry et al 2013/all-database-records.csv", 
             authentication = "2317c6d849b8c5c8e80a39691d544fe2",
             number_of_headers = "1",
             number_of_footers = "1",
             record_delimiter = "\\\\r\\\\n", 
             attribute_orientation = "column", 
             field_delimiter = ",")
}
