% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-element.R
\name{add_project}
\alias{add_project}
\title{Add Project}
\usage{
add_project(
  parent_element,
  title_metadata,
  project_personnel_metadata,
  funding_metadata
)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{title_metadata}{see \code{\link{create_project}}}

\item{project_personnel_metadata}{see \code{\link{create_project}}}

\item{funding_metadata}{see \code{\link{create_project}}}
}
\description{
Add Project
}
