% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-element.R
\name{add_project}
\alias{add_project}
\title{Add project element}
\usage{
add_project(
  parent_element,
  project_title,
  award_information,
  project_personnel
)
}
\arguments{
\item{parent_element}{A list in which the project node should be nested in}

\item{project_title}{The title of the project that the funding is awarded to.}

\item{award_information}{A list that includes that award title and the funder name
for the funding award.}

\item{project_personnel}{A list that includes the project personnel. It must include
the first name, last name, organization, and position for this project.}
}
\value{
This function returns the parent element with a new project node containing
all project information.
}
\description{
This function creates a project node within the parent element that
contains all the project information.
}
\examples{

add_project(project_title = "my project title", 
            award_information = list(title = "Money up for grabs",
                                     funderName = "Bank",
                                     awardNumber = "000",
                                     funderIdentifier = "Funder 1",
                                     awardUrl = "awardforme.com"),
            personnel_info = list(individualName = list(givenName = "Smithy",
                                                        surName ="Smith"),
                                  organizationName = "US GOV",
                                  electronicMailAddress = "myemail@mail.gov",
                                  role = "Manager"))
}
