% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-element.R
\name{add_project}
\alias{add_project}
\title{Add Project}
\usage{
add_project(
  parent_element,
  funding_metadata,
  project_title = NULL,
  project_lead = NULL
)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{funding_metadata}{Add a named list or dataframe containing funding metadata: see \code{\link{create_project}}}

\item{project_title}{Optionally add a project_title that is different from dataset title: see \code{\link{create_project}}}

\item{project_lead}{Optionally add metadata describing a specific project personnel: see \code{\link{create_project}}}
}
\value{
The dataset list with project information appended.
}
\description{
Adds the project metadata elements according to EML standards.
}
