% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-element.R
\name{add_project}
\alias{add_project}
\title{Add Project}
\usage{
add_project(
  parent_element,
  funding_metadata,
  project_title = NULL,
  project_lead = NULL
)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{funding_metadata}{Add a named list or dataframe containing funding metadata: see \code{\link{create_project}}}

\item{project_title}{Optionally add a project_title that is different from dataset title: see \code{\link{create_project}}}

\item{project_lead}{Optionally add metadata describing a specific project personnel: see \code{\link{create_project}}}
}
\value{
The dataset list with project information appended.
}
\description{
Adds the project metadata elements according to EML standards.
}
\examples{
project_lead <- dplyr::tibble(first_name = "Stacy", last_name = "Banet", email = "Stacy@aol.com" , 
                              role = "creator", organization = "USBR", orcid = NA)
funding_metadata <- list(funder_name = "USBR", funder_identifier = NA, award_number = "R14AC00096", 
                         award_title = "Salmonid Spawning and Rearing Habitat Restoration in the Sacramento River", 
                         award_url = NA, funding_description = NA)

dataset <- list() \%>\%
   add_title(list(title = "O.mykiss Habitat monitoring and contact point project", short_name = "O.mykiss monitoring")) \%>\%
   add_personnel(project_lead) \%>\%
   add_project(funding_metadata)
dataset

dataset <- list() \%>\%
   add_project(funding_metadata, project_title = "Salmonid Habitat monitoring in the Central Valley", project_lead = project_lead)   
dataset
}
