% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-element.R
\name{add_project}
\alias{add_project}
\title{Add project element}
\usage{
add_project(
  parent_element,
  project_title,
  award_information,
  project_personnel
)
}
\arguments{
\item{parent_element}{A list in which the project node should be nested in.}

\item{project_title}{The title of the project that the funding is awarded to.}

\item{award_information}{A list that includes the required funding information
for an EML document.This list must include the award title and the funderName.
This list can be created by calling the \code{add_funding} function
on the funding information or by manually inputting the required information.
If the list is written manually it must be formatted as follows.

\code{award_infomation = list(funderName = "Name", title = "Award Title")}

Additional information about the funding may be added to the list. See the
\code{\link{add_funding}} documentation for more information.}

\item{project_personnel}{A list that includes the required information on project
personnel for an EML document. It must include the first name, last name,
organization, and personnel role for this project. This list can be created
by calling the \code{add_personnel} function on the project personnel or by manually inputting the required
information. If the list is written manually it must be formatted as follows.

\code{project_personnel = list(individualName = list(givenName = "First Name", surName = "Last Name"),
role = "Position", organization = "Organization")}

Additional information about the project personnel may be added to the list. See the
\code{\link{add_personnel}} documentation for more information.}
}
\value{
This function returns the parent element with a new project node containing
all project information required for an EML document.
}
\description{
This function creates a project node within the parent element that
contains all the required elements for the project section of an EML document.
This function can be used in combination with \code{add_personnel}
and \code{add_funding}. \code{add_personnel} can be used to generate the \code{project_personnel}
input and \code{add_funding} can be used to generate the \code{award_information} input.
}
\examples{
add_project(parent_element = parent_element, 
            project_title = "my project title", 
            award_information = add_funding(funder_name = "Bank",
                                            funder_identifier = "Funder 1",
                                            award_number = "000",
                                            award_title = "Money up for grabs",
                                            award_url = "awardforme.com"),
            project_personnel = add_personnel(parent_element = parent_element,
                                              first_name = "Smithy",
                                              last_name = "Smith",
                                              email = "myemail@mail.gov",
                                              role = "Manager",
                                              organization = "US GOV"))
                                              
add_project(parent_element = parent_element, 
            project_title = "my project title", 
            award_information = list(funderName = "Bank",
                                     funderIdentifier = "Funder 1",
                                     awardNumber = "000",
                                     title = "Money up for grabs",
                                     awardUrl = "awardforme.com"),
            project_personnel = list(individualName = list(givenName = "Smithy",
                                                           surName = "Smith"),
                                     electronicMailAddress = "myemail@mail.gov",
                                     role = "Manager",
                                     organizationName = "US GOV"))
}
