% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-date-element.R
\name{add_pub_date}
\alias{add_pub_date}
\title{Add Pub Date}
\usage{
add_pub_date(parent_element, date = NULL)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{date}{The publication date. If left null, it will be assigned the current
date. This can be overridden by putting in a specific date: see \code{\link{create_pub_date}}}
}
\value{
The dataset list or project with publication date information appended.
}
\description{
Adds the pub date element to a dataset list according to EML standards.
}
\examples{
 
dataset <- list() \%>\%
    add_pub_date()
dataset
}
