% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-element.R
\name{add_raster}
\alias{add_raster}
\title{Add Raster File}
\usage{
add_raster(parent_element, raster_metadata)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{raster_metadata}{A named list or dataframe containing raster metadata elements
(file_name, file_description, attribute_list, physical, spatial_reference, horizontal_accuracy, vertical_accuracy,
cell_size_x, cell_size_y, number_of_bands, raster_origin, rows, columns, verticals, cell_geometry): see \code{\link{create_raster}}}
}
\value{
The dataset list or project with raster file information appended.
}
\description{
Adds the raster elements to a dataset list according to EML standards.
}
