% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic-coverage-element.R
\name{add_taxonomic_coverage}
\alias{add_taxonomic_coverage}
\title{Add Taxonomic coverage}
\usage{
add_taxonomic_coverage(
  parent_element,
  default_species = NULL,
  kingdom = "KINGDOM",
  kingdom_value = NULL,
  phylum = "PHYLUM",
  phylum_value = NULL,
  class = "CLASS",
  class_value = NULL,
  order = "ORDER",
  order_value = NULL,
  family = "FAMILY",
  family_value = NULL,
  genus = "GENUS",
  genus_value = NULL,
  species = "SPECIES",
  species_value = NULL,
  common_name = NULL
)
}
\arguments{
\item{default_species}{TODO}

\item{kingdom}{TODO}

\item{kingdom_value}{TODO}

\item{phylum}{TODO}

\item{phylum_value}{TODO}

\item{class}{TODO}

\item{class_value}{TODO}

\item{order}{TODO}

\item{order_value}{TODO}

\item{family}{TODO}

\item{family_value}{TODO}

\item{genus}{TODO}

\item{genus_value}{TODO}

\item{species}{TODO}

\item{species_value}{TODO}

\item{common_name}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}
