% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic-coverage-element.R
\name{add_taxonomic_coverage}
\alias{add_taxonomic_coverage}
\title{Add Taxonomic Coverage}
\usage{
add_taxonomic_coverage(taxonomic_metadata)
}
\arguments{
\item{taxonomic_metadata}{see \code{\link{create_taxonomic_coverage}}}
}
\description{
Add Taxonomic Coverage
}
