% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic-coverage-element.R
\name{add_taxonomic_coverage}
\alias{add_taxonomic_coverage}
\title{Add Taxonomic Coverage}
\usage{
add_taxonomic_coverage(
  CVPIA_common_species = NULL,
  kingdom = "kingdom",
  kingdom_value,
  phylum = "phylum",
  phylum_value,
  class = "class",
  class_value,
  order = "order",
  order_value,
  family = "family",
  family_value,
  genus = "genus",
  genus_value,
  species = "species",
  species_value,
  common_name
)
}
\arguments{
\item{CVPIA_common_species}{A list of common CVPIA species can be set as the default species so that taxonomic coverage does not need to be inputted in manually. These species include: chinook salmon, steelhead, delta smelt, white sturgeon and green sturgeon. If one of these species are applicable, use the CVPIA_common_species inputs of "chinook", "steelhead", "delta_smelt", "white_sturgeon", and "green_sturgeon".}

\item{kingdom}{Kingdom level present}

\item{kingdom_value}{The kingdom name.}

\item{phylum}{Phylum level present}

\item{phylum_value}{The phylum name.}

\item{class}{Class level present}

\item{class_value}{The class level name.}

\item{order}{Order level present}

\item{order_value}{The order level name.}

\item{family}{Family level present}

\item{family_value}{The family level name.}

\item{genus}{Genus level present}

\item{genus_value}{The genus level name.}

\item{species}{Species level present}

\item{species_value}{The species level name.}

\item{common_name}{The common name of the organism}
}
\value{
The dataset or project with taxonomic coverage information appended
}
\description{
Adds the taxonomic coverage information of a dataset based off of EML standards. The addition of taxonomic coverage is optional, however defaults of chinook, sturgeon, smelt, and steelhead are provided. While single or multiple taxonomies can be applied, the full set of information must be provided if chosen to be included.
}
\examples{
add_taxonomic_coverage(CVPIA_common_species = "chinook")

add_taxonomic_coverage(CVPIA_common_species = "smelt")

add_taxonomic_coverage(CVPIA_common_species = "sturgeon")

add_taxonomic_coverage(CVPIA_common_species = "steelhead")

add_taxonomic_coverage(kingdom_value = "Animalia",
                       phylum_value = "Chordata",
                       class_value = "Mammalia",
                       order_value = "Carnivora",
                       family_value = "Felidae",
                       genus_value = "Panthera", 
                       species_value = "Panthera Leo",
                       common_name = "Lion")
}
