% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic-coverage-element.R
\name{add_taxonomic_coverage}
\alias{add_taxonomic_coverage}
\title{Add Taxonomic Coverage}
\usage{
add_taxonomic_coverage(
  CVPIA_common_species = NULL,
  kingdom = "kingdom",
  kingdom_value,
  phylum = "phylum",
  phylum_value,
  class = "class",
  class_value,
  order = "order",
  order_value,
  family = "family",
  family_value,
  genus = "genus",
  genus_value,
  species = "species",
  species_value,
  common_name
)
}
\arguments{
\item{CVPIA_common_species}{Use one of the following: "chinook", "delta_smelt", "white_sturgeon", "green_sturgeon", or "steelhead" to get pre-selected information from ITIS}

\item{kingdom}{Kingdom level present}

\item{kingdom_value}{The kingdom name.}

\item{phylum}{Phylum level present}

\item{phylum_value}{The phylum name.}

\item{class}{Class level present}

\item{class_value}{The class level name.}

\item{order}{Order level present}

\item{order_value}{The order level name.}

\item{family}{Family level present}

\item{family_value}{The family level name.}

\item{genus}{Genus level present}

\item{genus_value}{The genus level name.}

\item{species}{Species level present}

\item{species_value}{The species level name.}

\item{common_name}{The common name of the organism}
}
\value{
Taxonomic coverage information. The function should be assigned to the name taxonomic_coverage to append it to the dataset or project. Example of how to incorporate it into the add_coverage function is seen below.
}
\description{
Adds the taxonomic coverage information of a dataset based off of EML standards. The addition of taxonomic coverage is optional, however defaults of chinook, sturgeon, smelt, and steelhead are provided. While single or multiple taxonomies can be applied, the full set of information must be provided if chosen to be included.
}
\section{CVPIA Common Species}{

By using a CVPIA common species, the appropriate taxonomy is appended based off of the Integrated Taxonomic Information System (ITIS).

Select "chinook" to append the taxonomy of Oncorhynchus tshawytscha (Also known as Chinook Salmon or King Salmon) from the ITIS database. For further taxonomic coverage on the Chinook Salmon, you can visit: \href{https://www.itis.gov/servlet/SingleRpt/SingleRpt?search_topic=TSN&search_value=161980#null}{ITIS's} webpage with full coverage.

Hi
}

\examples{
taxonomic_coverage <- add_taxonomic_coverage(CVPIA_common_species = "chinook")

taxonomic_coverage <- add_taxonomic_coverage(CVPIA_common_species = "delta_smelt")

taxonomic_coverage <- add_taxonomic_coverage(CVPIA_common_species = "white_sturgeon")

taxonomic_coverage <- add_taxonomic_coverage(CVPIA_common_species = "green_sturgeon")

taxonomic_coverage <- add_taxonomic_coverage(CVPIA_common_species = "steelhead")

taxonomic_coverage <- add_taxonomic_coverage(kingdom_value = "Animalia",
                       phylum_value = "Chordata",
                       class_value = "Mammalia",
                       order_value = "Carnivora",
                       family_value = "Felidae",
                       genus_value = "Panthera", 
                       species_value = "Panthera Leo",
                       common_name = "Lion")   
                                            
To append this information to the dataset or project:                        
    add_coverage(parent_element = list(), geographic_description = "Description",
}
