% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-element.R
\name{add_title}
\alias{add_title}
\title{Add Title Elements}
\usage{
add_title(parent_element, title, short_name)
}
\arguments{
\item{title}{The title of the project and/or dataset. A complete title is between
7 and 20 words long and includes: What, Where, and When (and Who, if relevant).}

\item{short_name}{short name of the project/dataset TODO add EDI recommendations}

\item{parament_element}{dataset or project}
}
\description{
Add Title Elements
}
