% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-element.R
\name{add_title}
\alias{add_title}
\title{Add Title}
\usage{
add_title(parent_element, title_metadata)
}
\arguments{
\item{parent_element}{A list representing the EML project or dataset.}

\item{title_metadata}{A list or dataframe containing a title and shortname: see \code{\link{create_title}}}
}
\description{
Add Title
}
\examples{
title_metadata <- list(title = "Eight Mile Lake Research Watershed, Carbon in Permafrost Experimental Heating Research.",
                       short_name = "knb-lter-bnz.501.17")
dataset <-list() \%>\% 
   add_title(title_metadata)
}
