% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-element.R
\name{create_project}
\alias{create_project}
\title{Create project element}
\usage{
create_project(project_title, project_lead, funding_metadata)
}
\arguments{
\item{project_title}{A project title as a string}

\item{project_lead}{A named list or datatable describing the project lead for the project. See structure by looking at required inputs of \code{\link{create_person}}}

\item{funding_metadata}{A named list or datatable describing the project funding. See structure by looking at required inputs of \code{\link{create_funding}}}
}
\value{
This function returns project element containing
all project information required for an EML document.
}
\description{
This function creates a project element from metadata inputs using  \code{create_person}
and \code{create_funding}.
}
\examples{
project_title <- "Salmonid Habitat Monitoring"
project_lead <- list(first_name = "Stacy", last_name = "Banet", email = "Stacy@aol.com", 
                                   role = "creator", organization = "USBR")
funding_metadata <- list(funder_name = "USBR", funder_identifier = NA, award_number = "R14AC00096", 
                         award_title = "Salmonid Spawning and Rearing Habitat Restoration in the Sacramento River", 
                         award_url = NA)

create_project(project_title, project_lead, funding_metadata)
}
