% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-element.R
\name{create_project}
\alias{create_project}
\title{Create project element}
\usage{
create_project(project_title, project_lead, funding_metadata)
}
\arguments{
\item{project_title}{A project title as a string}

\item{project_lead}{A named list or datatable describing the project lead for the project. See structure by looking at required inputs of \code{\link{create_person}}}

\item{funding_metadata}{A named list or datatable describing the project funding. See structure by looking at required inputs of \code{\link{create_funding}}}
}
\value{
This function returns project element containing
all project information required for an EML document.
}
\description{
This function creates a project element from metadata inputs using  \code{add_personnel}
and \code{add_funding}.
}
