% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-element.R
\name{create_vector}
\alias{create_vector}
\title{Create Vector Element}
\usage{
create_vector(file_name, file_description, attribute_info, physical, geometry)
}
\arguments{
\item{file_name}{The name of your vector file}

\item{file_description}{A short description of your file}

\item{attribute_info}{File path to metadata excel that contains attribute
information for vector data}

\item{physical}{A list of the physical descriptions of your file. Use
\code{\link{create_physical}} to generate this physical list and see the
documentation for \code{\link{create_physical}} for more information.}

\item{geometry}{Geometric representation of the file}

\item{parent_element}{A list to append \code{spatialVector} to}
}
\value{
A list that contains all the required elements of the spatialVector
section of an EML document.
}
\description{
This function creates all the required elements for a \code{spatialVector} section of
an ELM document.
}
\details{
Create Vector Element
}
\examples{
\dontrun{
create_vector(file_name = "vectorfiles.tif" ,
              file_description = "A vector File",
              attribute_info =  "metadata_vectorfiles.xlsx",
              physical = create_physical("vectorfiles.tif"),
              geometry = "pixel")}

}
