% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CVPIA_common_species}
\alias{CVPIA_common_species}
\title{CVPIA Common Species}
\format{
A named list, keys are names of CVPIA common species and values are nested
lists containing complete taxonomic information for each species.
}
\usage{
CVPIA_common_species
}
\description{
A named list containing the names and complete taxonomic information of
CVPIA common species. \code{CVPIA_common_species} is used by \code{add_taxonomic_coverage}
to generate the taxonomic coverage element according the the EML schema.
}
\section{CVPIA Common Species}{

\itemize{
\item \strong{chinook}
\item \strong{steelhead}
\item \strong{delta_smelt}
\item \strong{white_sturgeon}
\item \strong{green_sturgeon}
}
}

\examples{
CVPIA_common_species$chinook 

$taxonomicClassification
$taxonomicClassification$taxonRankName
[1] "kingdom"

$taxonomicClassification$taxonRankValue
[1] "Animalia"

$taxonomicClassification$taxonomicClassification
$taxonomicClassification$taxonomicClassification$taxonRankName
[1] "phylum"

$taxonomicClassification$taxonomicClassification$taxonRankValue
[1] "Chordata"

$taxonomicClassification$taxonomicClassification$taxonomicClassification
$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
[1] "class"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
[1] "Teleostei"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
[1] "order"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
[1] "Salmoniformes"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
[1] "family"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
[1] "Salmonidae"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
[1] "genus"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
[1] "Oncorhynchus"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification
$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankName
[1] "species"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonRankValue
[1] "Oncorhynchus tshawytscha"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$commonName
[1] "Chinook Salmon"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId
$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId$provider
[1] "https://itis.gov"

$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonomicClassification$taxonId$taxonId
[1] "161980"

}
\keyword{datasets}
