% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CVPIA_common_species}
\alias{CVPIA_common_species}
\title{CVPIA Common Species}
\format{
A named list, keys are named the same as values.
}
\usage{
CVPIA_common_species
}
\description{
A list of CVPIA common species that can be appended to the taxonomic
coverage.
}
\section{CVPIA Common Species}{

\itemize{
\item \strong{chinook}
\item \strong{steelhead}
\item \strong{delta_smelt}
\item \strong{white_sturgeon}
\item \strong{green_sturgeon}
}
}

\examples{
CVPIA_common_species$chinook #"chinook"

}
\keyword{datasets}
