% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CVPIA_funders}
\alias{CVPIA_funders}
\title{CVPIA_funders}
\format{
A named list, keys are names of CVPIA funding organizations and keys contain
information about each funder.
}
\usage{
CVPIA_funders
}
\description{
A list of CVPIA_funders that can be appended to the funding element.
}
\section{CVPIA_funders}{

\itemize{
\item \strong{USBR}
\item \strong{CDWR}
\item \strong{CDFW}
}
}

\examples{
CVPIA_funders$USBR # $funderName
                   #   [1] "United States Bureau of Reclamation"

                   # $funderIdentifier
                   #   [1] "https://www.wikidata.org/wiki/Q1010548"


}
\keyword{datasets}
